### 2023-09-17 #608 @csaybar
rgeeExtra:  An extensions for 'rgee'
Submitting Author Name: Cesar Aybar
Submitting Author Github Handle: @csaybar
Other Package Authors Github handles: (comma separated, delete if none) @davemlz, @LBautistaB13, @choisy
Repository:  https://github.com/r-earthengine/rgeeExtra
Version submitted: 0.0.1
Submission type: Standard

Version accepted: TBD

---

```
Package: rgeeExtra
Title: Extensions for 'rgee'
Version: 0.0.1
    c(person(given = "Cesar",
           family = "Aybar",
           email = "csaybar@gmail.com",
      person(given = "David",
           family = "Montero",
	         role = c("ctb"),
	         email = "dml.mont@gmail.com",
      person(given = "Lesly",
           family = "Bautista",
	         role = c("ctb"),
	         email = "leslyarcelly.213@gmail.com",
      person(given = "Marc",
           family = "Choisy",
	         role = c("ctb"),
	         email = "mchoisy@oucru.org",
    )
Description: High-level API to process Google Earth Engine (GEE) raster (ee.Image and ee.ImageCollection) and vector data (ee.Geometry, ee.Feature, and ee.FeatureCollection). Popular Third-party GEE algorithms are re-coded from Javascript and Python to R.
License: Apache License (>= 2)
Config/reticulate:
  list(
    packages = list(
      list(package = "earthengine-api")
    )
  )
    reticulate (>= 1.31),
    rgee,
    crayon,
    cli
    googledrive,
    googleCloudStorageR,
    geojsonio,
    magick,
    stars,
    sf,
```

[rgeeExtra](https://github.com/r-earthengine/rgeeExtra) is an extension of the [rgee](https://github.com/r-earthengine/rgee) package that enables interaction with [Google Earth Engine](https://earthengine.google.com/) (GEE). The advantage of rgeeExtra lies in its user-friendly syntax for R users. The package includes operator overloading and access to various functions created by third parties within the [ee_extra](https://github.com/r-earthengine/ee_extra) package.

GEE users and any R user who wants to retrieve data using the GEE API from within R.

There are other packages for connecting to GEE, but they are primarily focused on composite generation and data download. See [geedim](https://github.com/leftfield-geospatial/geedim)

Yes

-  Package uses global assignment operator ('<<-'). The global assignment operator is used in the .onLoad special function to set the *ee_extra* Python Package in the R global environment (See [here](https://github.com/r-earthengine/rgeeExtra/blob/master/R/rgeeExtra-package.R)). This approach is similar to other reticulate R packages like keras or tensorflow.
