### 2023-09-19 #608 @maurolepore
@csaybar thanks a lot for your submission.

* Can you confirm you're requesting a review in English?
* Also, this is part of the Champions program, right?

`@mpadge` kindly noted that coverage and R CMD check failed due to a missing python dependency that our system can't track. I'll need a bit of time to investigate more. Do you know a good/easy way to provide the required Python environment so that the handling editor and reviewers can hit the ground running?
