### 2023-09-21 #608 @csaybar
Hi @maurolepore thank you for your quick response. I just activated the notification button to this issue.

*Can you confirm you're requesting a review in English?*

Yes, I'd prefer to do it in English

Also, this is part of the Champions program, right?

Yes, I'm part of the program and my mentor is Marc Choisy @choisy.

Unfortunately, I haven't been able to resolve the issues related to Python dependencies. These are the same ones you can find in the [pangoling](https://github.com/bnicenboim/pangoling) package review. The Python coverage when it is evaluated in an ENV with the Python envs is around 95%. See https://app.codecov.io/gh/csaybar/rgeeExtra/tree/master/R
