+++ NEWS.md
@@ -1,3 +1,8 @@
+# rgeeExtra 0.0.1.999
+
+* Working in `pkgcheck` flags.
+* The flag "The following functions have no documented return values:[ee_length_ic, ee_length, ee_name_ic, ee_name, ee_subsetting, extra_Initialize, Math-methods, Ops-methods, Summary-methods]" has been solved.
+
 # rgeeExtra 0.0.1
 
 * First version of rgeeExtra
