### 2023-09-21 #608 @maurolepore
Great, thanks for pointing me to pangoling. I'll explore how they handled installation during [review](https://github.com/ropensci/software-review/issues/575).
