### 2023-09-22 #608 @maurolepore
@csaybar, I self-assigned to be the handling editor. I think this makes sense because I would like to explore the installation process, I'm part of the Champions program, and I'm available.

I'll start working on the handling-editor's checklist. In the meantime, could you please list three potential reviewers?

Following our guidelines I'll use no more than one of them, but it's useful to  have a few options and also to see the range of expertise you think would add value to this submission.
