### 2023-09-26 #608 @maurolepore
@csaybar,

In the installation instructions of rgee I see you point to a [posit.cloud project](https://posit.cloud/content/5175749) with a file that aims to quickly help a new user set everything up in a fresh environment. That seems like a good idea. Could you please do ti also for rgeeExtra, so that I and the reviewers can install the package as quickly as possible?

Note however that instructions for rgee don't seem to just work on posit.cloud as I seem to need `gcloud` and I can't use `sudo` to install it on posit.cloud. [Here is the project I used](https://posit.cloud/content/6633813). You'll see some FIXME that explain my struggles. As rgeeExtra depends on rgee I guess simplifying the installation of rgee will also help with rgeeExtra?

If posit.cloud limits the system dependencies we can install, how else can we quickly create a working environment for the reviewers? Would it help to have a Docker image on r-hub or a Dockerfile in the repo?
