### 2023-09-30 #608 @csaybar
Hi @maurolepore

Here is an example of how to use rgeeExtra in posit cloud.
https://posit.cloud/content/6670766

It is important that the rgeeExtra user has a **GEE account**.

Unfortunately, positCloud does not have a way to send the token back (I tried to solve this problem a year ago). Kevin Ushey, explain this [here](https://twitter.com/csaybar/status/1576427594383519744)

"It's an unfortunate limitation of system() in RStudio; we currently don't have a way of routing stdin from the R process to subprocesses. You could instead run this in a separate Terminal instance for now."

![image](https://github.com/ropensci/software-review/assets/16768318/441216d4-25d0-42c1-831f-580bc91b6de6)

So the rgee user who wants to use PositCloud should do so through the terminal using:

```
$ earthengine authenticate --auth_mode notebook
```

![image](https://github.com/ropensci/software-review/assets/16768318/df37597e-6a8d-4f3e-8474-55c0d88f34ec)

After that rgee & rgeeExtral will work perfectly! I attached two examples that show rgeeExtra features.

![image](https://github.com/ropensci/software-review/assets/16768318/7f27f0d8-4407-4791-90ab-abb7500bb8fd)

Feel free to reach out if there's anything I can assist you with!
