### 2023-10-10 #608 @maurolepore
Thanks @csaybar,

I tried your instructions for installation and usage a few times. I arrived to the version below that seems succinct and functional. What do you think?

After your review it would be nice to edit README to ensure the smoothest experience for anyone trying to install and use the package.

----

## Installation

```r
install.packages("remotes")
remotes::install_github("r-spatial/rgee")
remotes::install_github("r-earthengine/rgeeExtra")
rgee::ee_install()
```

Authenticate with Google Earth Engine

```bash
# Run in the Terminal (not in the R console)
earthengine authenticate --auth_mode notebook
```

## Example

```r
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

img <- ee$Image("LANDSAT/LC08/C01/T1_SR/LC08_038029_20180810")
ndvi <- (img[["B5"]] - img[["B4"]])/(img[["B5"]] + img[["B4"]])**2
names(ndvi) <- "pow_ndvi"

Map$centerObject(img,zoom=12)
Map$addLayer(img)
```

---

Now that I can install the package I will continue with all other editor checks.
