+++ README.md
@@ -27,7 +27,6 @@ developed."></a>
 status"></a>
 <a href="https://doi.org/10.5281/zenodo.3945409"><img src="https://zenodo.org/badge/DOI/10.5281/zenodo.3945409.svg" alt="DOI"></a>
 <br>
-<a href="https://github.com/r-earthengine/rgeeExtra/actions/workflows/R-CMD-check.yaml"><img src="https://github.com/r-earthengine/rgeeExtra/actions/workflows/R-CMD-check.yaml/badge.svg" alt="" width="200"></a>
 <a href="https://www.buymeacoffee.com/csay" target="_blank"><img src="https://www.buymeacoffee.com/assets/img/custom_images/orange_img.png" alt="Buy Me A Coffee" style="height: 41px !important;width: 174px !important;box-shadow: 0px 3px 2px 0px rgba(190, 190, 190, 0.5) !important;-webkit-box-shadow: 0px 3px 2px 0px rgba(190, 190, 190, 0.5) !important;" ></a>
 </p>
 
@@ -102,6 +101,9 @@ You can install rgeeExtra from [GitHub](https://github.com/r-earthengine/rgeeExt
 ``` r
 remotes::install_github("r-earthengine/rgeeExtra")
 ```
+
+To install Python dependencies, please refer to the installation guidelines provided in the [rgee](https://github.com/r-spatial/rgee#installation) 
+section of the documentation.
 
 ## How does it work?
 
