+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   Features: Features.html
   Introduction: Introduction.html
-last_built: 2023-09-21T13:19Z
+last_built: 2023-10-15T09:58Z
 urls:
   reference: https://r-earthengine.github.io/rgeeExtra/reference
   article: https://r-earthengine.github.io/rgeeExtra/articles
