### 2023-10-15 #608 @maurolepore
---

@csaybar,

To help you track my comments, I'll label them with "ml" and a numerical sequence: ml01, ml02, etc:

* Comments starting with "bullet" (like this one) suggest things that do not require your action or response.

- [ ]  Comments starting with "checkbox" (like this one) ask for things that require your action and response.

--

**The package has sufficient documentation available online (README, pkgdown docs) to allow for an assessment of functionality and scope without installing the package.**

- [ ] ml01. In all code examples please show not only the code but also its outout. This will help the reviewers and users to assess the package functionality without installing the package and running the code.

You may evaluate code chunks in .Rmd or .qmd documents or with the [reprex package](https://reprex.tidyverse.org/). It's best to use the option `comment = "#>"` so that the motivated reader can copy/paste/run without errors.

**Is the case for the package well made?**

- [ ] ml02. Please Answer "Why rgeeExtra is needed?" at the top of README. That location is the most useful for reviewers and new users.

Note the [output of `usethis::use_readme_rmd()`](https://github.com/r-lib/usethis/blob/main/inst/templates/package-README) starts with "The goal of {{{ Package }}} is to ...".

Currently the top of README talks about GEE and rgee but not rgeeExtra ([What is Google Earth Engine?](https://github.com/r-earthengine/rgeeExtra#what-is-google-earth-engine)). The reader sees the [Installation or rgeeExtra](https://github.com/r-earthengine/rgeeExtra#installation) before knowing about it.

I think the best case for the package is made at top of the [Features article](https://r-earthengine.com/rgeeExtra/articles/Features.html). Moving this text to the top or README would solve the problem :-)

> rgeeExtra serves as a wrapper for the Python package named eeExtra. The creation of eeExtra was driven by a need to consolidate various third-party GEE packages and projects found on GitHub.

> rgeeExtra ensures a seamless integration of these projects across different programming languages by standardizing methods and facilitating the use of JavaScript code projects beyond the traditional Code Editor environment.

Also in other places I see text that can help make the case for the package:

https://r-earthengine.com/rgeeExtra/articles/Introduction.html:

> rgeeExtra is an extension of the rgee package, enhancing the capabilities of Google Earth Engine (GEE) integration within the R environment.

https://r-earthengine.com/rgeeExtra/index.html#how-does-it-work:

> The rgeeExtra extends the following Earth Engine classes: ...

In any case, all candidate text seems to answer "what is rgeeExtra" but to make a stronger case for the package the question that needs to be answered is "why is reeExtra needed?".

**Is the reference index page clear (grouped by topic if necessary)?**

- [ ] ml03. In the [reference index](https://r-earthengine.com/rgeeExtra/reference/index.html), please [group the functions by topic](https://devdevguide.netlify.app/pkg_building#function-grouping). This will help the reviewers and users to immediately see the structure of the public API. It may also help you find gaps or redundancy.

**Are vignettes readable, sufficiently detailed and not just perfunctory?**

* ml04. The content of the articles seems very useful, and relocating some of it can help to ensure the content reaches reviewers and users where they most commonly need it and expect it. For example:

    - Make the case for the package at the top of README (moving text from the article "Features" as suggested above).
    - Show installation only in README, before the first usage of the package, and ask multiple new users to confirm they can install and use the package by following the instructions in the exact order they are presented in the installation section in README. Also ensure they follow and give feedback on instructions linked to from README but outside of it.
    - Turn the article [Introduction](https://r-earthengine.com/rgeeExtra/articles/Introduction.html) into the common "Get started" article (via `usethis::use_article("name-of-your-package")`). Remove the installation section and instead point to the installation in README. If you keep installation instructions, ensure they come before (not after) the first example of usage. Considering explaining the code in prose step by step, rather than showing a big code-chunk then explaining it all after the end. The final figure seems misplaced. It would be more natural to see a plot as the output of a code chunk -- so it should follow the code, not the prose.

**Fit: The package meets criteria for fit and overlap.**

* ml05. The package fits nicely in the stated categories, but consider explaining more about the overlap with rgee. For inspiration see how the [targets package explains why it supersedes drake](https://books.ropensci.org/targets/drake.html).

**Installation instructions: Are installation instructions clear enough for human users?**

- [ ] mlxx. As is the installation instructions in README didn't work for me. After several iterations I arrived to
a this revision: https://github.com/ropensci/software-review/issues/608#issuecomment-1755823040.
I suggest asking other newbies like me to try it before asking the reviewers to install it. The limited reviewer's time and
energe will be better spent in providing more interesting feedback.

**Tests: If the package has some interactivity / HTTP / plot production etc. are the tests using state-of-the-art tooling?

- [ ] ml06. As much as possible please move tests setup code from each test file to a single [setup.R file](https://r-pkgs.org/testing-design.html#testthat-setup-files), a test helper, or other [advanced testing techniques](https://r-pkgs.org/testing-advanced.html).

- [ ] ml07. Please help me understand why there are many more functions in the [reference index](https://r-earthengine.com/rgeeExtra/reference/index.html) than test-files? It's common for each `R/exported_function.R` to have a corresponding `test/testthat/test-exported_function.R`.

- [ ] ml08. Please address the R CMD check failure on GitHub. The last passing commit dates August 18, 2023: https://github.com/r-earthengine/rgeeExtra/commits/master. Do you need to fix some test, or maybe use `skip_on_ci()` or similar?

**Contributing Is: information the documentation for contribution clear enough e.g. tokens for tests, playgrounds?**

- [ ] ml09. Is there any place in the docs where you explain how a contributor or user gets an account with GEE and a token? If not, please add it to the installation instrucitons or wherever you think it's most needed.

**License:** The package has a CRAN or OSI accepted license.

Yes.

**Project management: Are the issue and PR trackers in a good shape, e.g. are there outstanding bugs, is it clear when feature requests are meant to be tackled?**

- [ ] ml10. Please review the open issues. The total number of issues is fairly small, and it seems you may be able to close some already (e.g. 4, 5). For the remaining open issues, it would be nice to see a comment where you explain if you plan to address them. It's important to show the reviewers that we've made as much progress on our own before asking their feedback.

**Also**

* Running the exmple in README throws:

```r
#> The legacy packages maptools, rgdal, and rgeos, underpinning the sp package,
#> which was just loaded, were retired in October 2023.
#> Please refer to R-spatial evolution reports for details, especially
#> https://r-spatial.org/r/2023/05/15/evolution4.html.
#> It may be desirable to make the sf package available;
#> package maintainers should consider adding sf to Suggests:.
```

 I'm puzzled. Both rgee and rgeeExtra are alraedy using sf. Do you know where this is comming from?
