+++ inst/ee_extra/ee_extra/__init__.py
@@ -6,19 +6,6 @@ __version__ = "0.0.15"
 
 from ee_extra.Image.basic import *
 from ee_extra.ImageCollection.core import *
-from ee_extra.JavaScript import (
-    translate_general,
-    translate_utils,
-    translate_functions,
-    translate_jsm_extra,
-    translate_jsm_wrappers,
-    translate_specialfunctions,
-    translate_loops,
-)
-from ee_extra.JavaScript import translate_jsm_main
-from ee_extra.JavaScript.translate_main import translate
-from ee_extra.JavaScript.install import install
-from ee_extra.JavaScript.merge import require
 from ee_extra.QA import clouds, pipelines
 from ee_extra.Spectral import core
 from ee_extra.STAC import core
