+++ README.md
@@ -1,176 +1,166 @@
+---
+editor_options: 
+  markdown: 
+    wrap: 72
+---
+
 <h1 align="center">
-  <br>
-  <a href="https://r-spatial.github.io/rgee/"><img src="https://user-images.githubusercontent.com/16768318/118376965-5f7dca80-b5cb-11eb-9a82-47876680a3e6.png" alt="" width="200"></a>
-  <a href="https://r-earthengine.com/rgeeExtra/"><img src="https://user-images.githubusercontent.com/16768318/118376968-63a9e800-b5cb-11eb-83e7-3f36299e17cb.png" alt="" width="200"></a>
-  <a href="https://r-earthengine.com/rgeebook/"><img src="https://user-images.githubusercontent.com/16768318/118376966-60aef780-b5cb-11eb-8df2-ca70dcfe04c5.png" alt="" width="200"></a> 
-  <br>
-  rgeeExtra: An Extension for rgee
-  <br>
+
+<br>
+<a href="https://r-spatial.github.io/rgee/"><img src="https://user-images.githubusercontent.com/16768318/118376965-5f7dca80-b5cb-11eb-9a82-47876680a3e6.png" width="200"/></a>
+<a href="https://r-earthengine.com/rgeeExtra/"><img src="https://user-images.githubusercontent.com/16768318/118376968-63a9e800-b5cb-11eb-83e7-3f36299e17cb.png" width="200"/></a>
+<a href="https://r-earthengine.com/rgeebook/"><img src="https://user-images.githubusercontent.com/16768318/118376966-60aef780-b5cb-11eb-8df2-ca70dcfe04c5.png" width="200"/></a>
+<br> rgeeExtra: An Extension for rgee <br>
+
 </h1>
 
 <h4 align="center">
-Simplifies the interaction with GEE APIs and functions, making it more R-like. Popular third-party GEE algorithms are 
-made available to R users
+
+Simplifies the interaction with GEE APIs and functions, making it more
+R-like. Popular third-party GEE algorithms are made available to R users
+
 </h4>
 
-
 <p align="center">
-<a href="https://colab.research.google.com/github/r-spatial/rgee/blob/examples/rgee_colab.ipynb"><img src="https://colab.research.google.com/assets/colab-badge.svg" alt="Open in Colab" title="Open and Execute in Google Colaboratory"></a>
+
+<a href="https://colab.research.google.com/github/r-spatial/rgee/blob/examples/rgee_colab.ipynb"><img src="https://colab.research.google.com/assets/colab-badge.svg" title="Open and Execute in Google Colaboratory" alt="Open in Colab"/></a>
 <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project Status: Active – The project has reached a stable, usable
 state and is being actively
-developed."></a>
-<a href="https://codecov.io/gh/csaybar/rgeeExtra">
-  <img src="https://codecov.io/gh/csaybar/rgeeExtra/branch/master/graph/badge.svg"/>
+developed."/></a> <a href="https://codecov.io/gh/csaybar/rgeeExtra">
+<img src="https://codecov.io/gh/csaybar/rgeeExtra/branch/master/graph/badge.svg"/>
 </a>
-<a href="https://opensource.org/licenses/Apache-2.0"><img src="https://img.shields.io/badge/License-Apache%202.0-blue.svg" alt="License"></a>
-<a href="https://www.tidyverse.org/lifecycle/#maturing"><img src="https://img.shields.io/badge/lifecycle-maturing-blue.svg" alt="lifecycle"></a>
[…]
