+++ man/ee-cash-Image-cash-Extra_getCitation-open-paren-x-close-paren.Rd
@@ -1,14 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ee_Image.R, R/ee_ImageColletion.R
-\name{ee-citation}
-\alias{ee-citation}
+% Please edit documentation in R/ee_Image.R
+\name{ee$Image$Extra_getCitation(x)}
+\alias{ee$Image$Extra_getCitation(x)}
 \alias{ee_image_getCitation}
-\alias{ee_ImageCollection_getCitation}
 \title{Citing EE objects in publications}
 \usage{
-ee_image_getCitation(x)
-
-ee_ImageCollection_getCitation(x)
+`ee$Image$Extra_getCitation()`
 }
 \arguments{
 \item{x}{An EE object to get the citation from.}
@@ -27,6 +24,7 @@ ee_Initialize()
 
 ee$ImageCollection("NASA/GPM_L3/IMERG_V06") \%>\%
   ee$ImageCollection$first() \%>\%
-  ee$Image$getCitation()
+  ee$Image$Extra_getCitation()
 }
 }
+\concept{ee_Image}

[…]
