+++ R/ee_Image.R
@@ -3,10 +3,8 @@
 #' If it exists, retrieve the citation of an EE object.
 #'
 #' @param x An EE object to get the citation from.
-#'
-#' @name ee$Image$Extra_getCitation(x)
-#' @usage `ee$Image$Extra_getCitation()`
-#' @family ee_Image
+#' @name ee-citation
+#' @usage `ee$Image$Extra_getCitation(x)`
 #' @returns A character with citation information.
 #'
 #' @examples
@@ -20,7 +18,6 @@
 #'   ee$Image$Extra_getCitation()
 #' }
 
-
 ee_image_getCitation <- function(x) {
   EEextra_PYTHON_PACKAGE <- load_ee_Extra()
   EEextra_PYTHON_PACKAGE$STAC$core$getCitation(x = x)
@@ -32,7 +29,7 @@ ee_image_getCitation <- function(x) {
 #' If it exists, retrieve the DOI of an EE object.
 #'
 #' @name ee-getdoi
-#' @usage `ee$Image$Extra_getDOI()`
+#' @usage `ee$Image$Extra_getDOI(x)`
 #' @param x An EE object.
 #'
 #' @returns A character with DOI information.
@@ -62,7 +59,7 @@ ee_image_getDOI <- function(x) {
 #'
 #' @param x An ee$Image or an ee$ImageCollection object.
 #' @name ee-getoffset
-#' @usage `ee$Image$Extra_getOffsetParams()`
+#' @usage `ee$Image$Extra_getOffsetParams(x)`
 #' @returns A list with the offset parameters for each band.
 #'
 #' @examples
@@ -91,7 +88,7 @@ ee_Image_getOffsetParams <- function(x) {
 #'
 #' @param x An ee$Image.or an ee$ImageCollection object.
 #' @name ee-getscaleparams
-#' @usage `ee$Image$Extra_getScaleParams()`
+#' @usage `ee$Image$Extra_getScaleParams(x)`
 #' @returns A list with the scale parameters for each band.
 #' @examples
 #' \dontrun{
@@ -116,7 +113,7 @@ ee_Image_getScaleParams <- function(x) {
 #'
 #' @param x An ee$Image or an ee$ImageCollection object.
 #' @name ee-getstac
-#' @usage `ee$Image$Extra_getSTAC()`
+#' @usage `ee$Image$Extra_getSTAC(x)`
 #' @returns Return STAC metadata for each band.
 #' @examples
 #' \dontrun{
@@ -138,18 +135,25 @@ ee_Image_getSTAC <- function(x) {
 
 #' Adjust the image's histogram to match a target image
 #'
-#' @usage `ee$Image$Extra_matchHistogram()`
-#' @param image ee$Image to adjust.
-#' @param target ee$Image image to match.
-#' @param bands A dictionary of band names to match, with
-#' source bands as keys and target bands as values.
-#' @param geometry ee$Geometry. The region to match histograms in that
-#' overlaps both images. If none is provided, the geometry of the source
-#' image will be used.  Defaults is NULL.
-#' @param maxBuckets Integer. The maximum number of buckets to use when building
-#' histograms. Will be rounded to the nearest power of 2.  Default to 256.
+#' Matches the histogram of one image (source) to that of another image (target).
 #'
-#' @returns The adjusted image containing the matched source bands.
+#' @param x ee$Image to adjust.
+#' @param ... Additional arguments for histogram matching.
+#' See details for more information.
+#'
+#' @details
+#' The `...` argument can include the following:
+#' \itemize{
+#'   \item{target}{ee$Image. The target image to match.}
+#'   \item{bands}{Dictionary. Band names to match, with source bands as keys and target bands as values.}
[…]
