+++ tests/testthat/test-raster.R
@@ -8,14 +8,14 @@ extra_Initialize()
 
 img <- ee$ImageCollection$Dataset$LANDSAT_LC08_C01_T1$first()[["B1"]]
 
-test_that("ee_maxValue/ee_minValue", {
-  expect_type(ee_maxValue(img), "integer")
-  expect_type(ee_maxValue(img, "Points", sample_size = 1), "integer")
-  expect_type(ee_minValue(img), "integer")
-  expect_type(ee_minValue(img, "Points", sample_size = 1), "integer")
+test_that("ee$Image$Extra_maxValue/ee$Image$Extra_minValue", {
+  expect_type(ee$Image$Extra_maxValue(img), "integer")
+  expect_type(ee$Image$Extra_maxValue(img, "Points", sample_size = 1), "integer")
+  expect_type(ee$Image$Extra_minValue(img), "integer")
+  expect_type(ee$Image$Extra_minValue(img, "Points", sample_size = 1), "integer")
 
-  expect_error(ee_minValue(img, "points"))
-  expect_error(ee_maxValue(img, "points"))
+  expect_error(ee$Image$Extra_minValue(img, "points"))
+  expect_error(ee$Image$Extra_maxValue(img, "points"))
 })
 
 
