+++ tests/testthat/test-eeImageCollection.R
@@ -45,14 +45,6 @@ test_that("ee$ImageCollection$Extra_getSTAC", {
   expect_is(eelist, "list")
 })
 
-
-# test_that("ee_ImageCollection_spectralIndex", {
-#   s2_indices <- ee$ImageCollection("COPERNICUS/S2_SR") %>%
-#     ee$ImageCollection$Extra_preprocess() %>%
-#     ee$ImageCollection$Extra_spectralIndex(c("NDVI", "SAVI"))
-#   expect_is(s2_indices, "ee.imagecollection.ImageCollection")
-# })
-
 test_that("ee$ImageCollection$Extra_preprocess", {
   ee$ImageCollection("COPERNICUS/S2_SR") %>%
     ee$ImageCollection$Extra_preprocess() -> eeic
@@ -66,14 +58,3 @@ test_that("ee$ImageCollection$Extra_scaleAndOffset", {
   expect_is(eeic, "ee.imagecollection.ImageCollection")
 })
 
-# test_that("ee_ImageCollection_panSharpen", {
-#   img <- ee$Image("LANDSAT/LC08/C01/T1_TOA/LC08_047027_20160819")
-#   img_sharp <- ee$ImageCollection$Extra_panSharpen(img, method="HPFA", qa=c("MSE", "RMSE"), maxPixels=1e13)
-#   expect_is(img_sharp, "ee.image.Image")
-# })
-
-# test_that("ee_ImageCollection_tasseledCap", {
-#   img <- ee$Image("LANDSAT/LT05/C01/T1/LT05_044034_20081011")
-#   img <- ee$ImageCollection$Extra_tasseledCap(img)
-#   expect_is(img, "ee.image.Image")
-# })
