+++ R/ee_Image.R
@@ -9,13 +9,25 @@
 #'
 #' @examples
 #' \dontrun{
+#' library(rgee)
 #' library(rgeeExtra)
 #'
 #' ee_Initialize()
-#'
-#' ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
-#'   ee$ImageCollection$first() %>%
+#' extra_Initialize()
+#' 
+#' # Retrieve citation for the first image in NASA's IMERG V06 collection
+#' citation <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>% 
 #'   ee$Image$Extra_getCitation()
+#' 
+#' # Display the citation
+#' citation
+#' 
+#' # Fetching NASA/GPM_L3/IMERG_V06 image collection and retrieving its citation.
+#' citation_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
+#'   ee$ImageCollection$Extra_getCitation()
+#' 
+#' # Display the citation
+#' citation_
 #' }
 
 ee_image_getCitation <- function(x) {
@@ -30,19 +42,29 @@ ee_image_getCitation <- function(x) {
 #'
 #' @name ee-getdoi
 #' @usage `ee$Image$Extra_getDOI(x)`
-#' @param x An EE object.
+#' @param x An EE object to get the DOI from.
 #'
 #' @returns A character with DOI information.
 #'
 #' @examples
 #' \dontrun{
+#' library(rgee)
 #' library(rgeeExtra)
 #'
 #' ee_Initialize()
-#'
-#' ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
-#'   ee$ImageCollection$first() %>%
+#' extra_Initialize()
+#' 
+#' # Fetch DOI for first image in NASA IMERG V06 collection
+#' doi <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>% 
 #'   ee$Image$Extra_getDOI()
+#'
+#' doi 
+#' 
+#' # Retrieve and print the DOI for the NASA/GPM_L3/IMERG_V06 image collection.
+#' doi_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
+#'   ee$ImageCollection$Extra_getDOI()
+#' 
+#' doi_
 #' }
 ee_image_getDOI <- function(x) {
   EEextra_PYTHON_PACKAGE <- load_ee_Extra()
@@ -64,14 +86,25 @@ ee_image_getDOI <- function(x) {
 #'
 #' @examples
 #' \dontrun{
+#' library(rgee)
 #' library(rgeeExtra)
 #'
 #' ee_Initialize()
+#' extra_Initialize()
 #'
-#' ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
-#'   ee$ImageCollection$first() %>%
+#' # Retrieve offset parameters from the first image in NASA IMERG V06 collection
+#' offset_params <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>%
 #'   ee$Image$Extra_getOffsetParams()
+#' 
+#' # Display offset parameters for each band.
+#' offset_params
 #'
+#' # Get offset parameters from NASA IMERG V06 image collection.
+#' offset_params_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
+#'   ee$ImageCollection$Extra_getOffsetParams()
+#' 
+#' # Display offset parameters for each band.
+#' offset_params_
 #' }
 ee_Image_getOffsetParams <- function(x) {
   EEextra_PYTHON_PACKAGE <- load_ee_Extra()
[…]
