+++ man/ee_names_feature.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ops-subsetting-fc.R
+\name{ee_names_feature}
+\alias{ee_names_feature}
+\alias{names.ee.feature.Feature}
+\title{Names of Earth Engine FeatureCollection object}
+\usage{
+\method{names}{ee.feature.Feature}(x)
+}
+\arguments{
+\item{x}{an EE FeatureCollection object.}
+}
+\value{
+A vector representing the property names of the ee$FeatureCollection object
+}
+\description{
+Get the properties names of FeatureCollection object
+}
+\examples{
+\dontrun{
+library(rgee)
+library(rgeeExtra)
+
+ee_Initialize()
+extra_Initialize()
+
+fc <- ee$FeatureCollection('WRI/GPPD/power_plants')
+fc$propertyNames()$getInfo()
+}
+}
