+++ pkgdown.yml
@@ -2,9 +2,14 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
+  Aditional: Aditional.html
+  Feature: Feature.html
+  FeatureCollection: FeatureCollection.html
   Features: Features.html
+  Image: Image.html
+  ImageCollection: ImageCollection.html
   Introduction: Introduction.html
-last_built: 2023-11-09T11:07Z
+last_built: 2023-11-22T23:33Z
 urls:
   reference: https://r-earthengine.github.io/rgeeExtra/reference
   article: https://r-earthengine.github.io/rgeeExtra/articles
