+++ R/ee_Image.R
@@ -14,18 +14,18 @@
 #'
 #' ee_Initialize()
 #' extra_Initialize()
-#' 
+#'
 #' # Retrieve citation for the first image in NASA's IMERG V06 collection
-#' citation <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>% 
+#' citation <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>%
 #'   ee$Image$Extra_getCitation()
-#' 
+#'
 #' # Display the citation
 #' citation
-#' 
+#'
 #' # Fetching NASA/GPM_L3/IMERG_V06 image collection and retrieving its citation.
 #' citation_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
 #'   ee$ImageCollection$Extra_getCitation()
-#' 
+#'
 #' # Display the citation
 #' citation_
 #' }
@@ -53,17 +53,17 @@ ee_image_getCitation <- function(x) {
 #'
 #' ee_Initialize()
 #' extra_Initialize()
-#' 
+#'
 #' # Fetch DOI for first image in NASA IMERG V06 collection
-#' doi <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>% 
+#' doi <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>%
 #'   ee$Image$Extra_getDOI()
 #'
-#' doi 
-#' 
+#' doi
+#'
 #' # Retrieve and print the DOI for the NASA/GPM_L3/IMERG_V06 image collection.
 #' doi_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
 #'   ee$ImageCollection$Extra_getDOI()
-#' 
+#'
 #' doi_
 #' }
 ee_image_getDOI <- function(x) {
@@ -95,14 +95,14 @@ ee_image_getDOI <- function(x) {
 #' # Retrieve offset parameters from the first image in NASA IMERG V06 collection
 #' offset_params <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>%
 #'   ee$Image$Extra_getOffsetParams()
-#' 
+#'
 #' # Display offset parameters for each band.
 #' offset_params
 #'
 #' # Get offset parameters from NASA IMERG V06 image collection.
 #' offset_params_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
 #'   ee$ImageCollection$Extra_getOffsetParams()
-#' 
+#'
 #' # Display offset parameters for each band.
 #' offset_params_
 #' }
@@ -130,18 +130,18 @@ ee_Image_getOffsetParams <- function(x) {
 #'
 #' ee_Initialize()
 #' extra_Initialize()
-#' 
+#'
 #' # Retrieve scale parameters from the first image in NASA IMERG V06 collection
 #' scale_params <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>%
 #'   ee$Image$Extra_getScaleParams()
-#' 
+#'
 #' # Display scale parameters for each band in the image.
 #' scale_params
-#' 
+#'
 #' # Retrieve scale parameters for the NASA IMERG V06 collection.
 #' scale_params_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") %>%
 #'   ee$Image$Extra_getScaleParams()
-#' 
+#'
 #' # Display scale parameters for each band in the image.
 #' scale_params_
 #' }
@@ -170,13 +170,13 @@ ee_Image_getScaleParams <- function(x) {
 #' # Retrieve STAC metadata for the first image in NASA's GPM L3 IMERG V06 collection
[…]
