+++ tests/testthat.R
@@ -1,11 +1,12 @@
+# This file is part of the standard setup for testthat.
+# It is recommended that you do not modify it.
+#
+# Where should you do additional test configuration?
+# Learn more about the roles of various files in:
+# * https://r-pkgs.org/testing-design.html#sec-tests-files-overview
+# * https://testthat.r-lib.org/articles/special-files.html
+
 library(testthat)
-library(rgee)
 library(rgeeExtra)
-library(stars)
-library(sf)
 
-if (identical(Sys.getenv("NOT_CRAN"), "true")) {
-  test_check("rgeeExtra")
-}
-
-
+test_check("rgeeExtra")
