+++ tests/testthat/test-subsetting.R
@@ -1,5 +1,3 @@
-context("rgeeExtra: subsetting ee.Image")
-
 library(rgee)
 library(rgeeExtra)
 
@@ -19,7 +17,7 @@ test_that("ee_get ImageCollection", {
 
 test_that("ee_get FeatureCollection", {
   expect_error(rgeeExtra::ee_get(fc, -1))
-  expect_is(rgeeExtra::ee_get(fc, 0), "ee.featurecollection.FeatureCollection")
+  expect_s3_class(rgeeExtra::ee_get(fc, 0), "ee.featurecollection.FeatureCollection")
   expect_equal(fc$first()$getInfo(), ee_get(fc, 0:1)$first()$getInfo())
   expect_equal(fc$first()$getInfo(), ee_get(fc, c(0, 2))$first()$getInfo())
 })
