+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   Image: Image.html
   ImageCollection: ImageCollection.html
   Introduction: Introduction.html
-last_built: 2023-11-22T23:33Z
+last_built: 2023-11-23T11:56Z
 urls:
   reference: https://r-earthengine.github.io/rgeeExtra/reference
   article: https://r-earthengine.github.io/rgeeExtra/articles
