+++ .github/workflows/R-CMD-check.yaml
@@ -41,9 +41,9 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::rcmdcheck
-          needs: check
-
+          args: 'c("--no-manual", "--as-cran", "--no-tests")'
+          error-on: '"error"'
+          check-dir: '"check"'
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
