### 2023-11-28 #608 @maurolepore
Dear @csaybar,

I'm exploring the potential reviewers you listed [above](https://github.com/ropensci/software-review/issues/608#issuecomment-1731436320) and realize I forgot to highlight that our policies about [conflict of interest for reviewers](https://devdevguide.netlify.app/softwarereview_policies.html#coi) (COI) might shrink the list. For example, [David's profile](https://github.com/davemlz) suggest a COI: "I'm the ... co-creator of [eeExtra](https://github.com/r-earthengine/ee_extra) and [easystac](https://github.com/cloudsen12/easystac) (with the incredible [@csaybar](https://github.com/csaybar))". If that contribution was within the last year, then it would be considered a COI.

Can you please review the list with COI in mind? And while you're there, please briefly explain why you would value the revew of each one in your list? This is useful for me to sense what kind of experience you're looking for, so I can search for a similar profile elsewhere.

And if your list above already considers COI then great. I just need to confirm.
