### 2023-11-28 #608 @csaybar
Hi @maurolepore, my apologies for the confusion. I have updated the list in accordance with the COI and added a brief explanation. Please let me know if you have any questions or need further clarification. Thank you!
