### 2023-12-18 #608 @maurolepore
Hola @mdelfin,

Muchas gracias por tu interes in rgeeExtra y por tu feedback. Nota que esta pagina esta dedicada al proceso de revision de rOpenSci.

Quizas estas buscando esta otra pagina?

https://github.com/r-earthengine/rgeeExtra/issues
