### 2023-12-18 #608 @mdelfin
Hola Mauro, la verdad estoy haciendo una app en shiny y con conexion a google earth engine con la liberia rgee, lo que mencionan en la parte de los comentarios, es la conexion con POSIT CLOUD, es ahi donde tengo el problema, no puedo conectar, ya que los pasos que mencionan de la termninal, esto esta todo muy bien, PERO al momento de relalizar el ejercicion con  ee_Initialize(), vuelve el mismo problema que menciona ustedes en la parte de arriba. Saludos
