### 2024-01-06 #608 @zackarno
Thanks for the fantastic package and the chance to review! I look forward to incorporating it into my work and exploring more use-cases over time!

I attempted to limit the scope of my review to just `{rgeeExtra}`, avoid  `{rgee}` functionality, and also not dig into the `EE_extra`/`EE-Mont` python packages which are wrapped.

Below is my review, I tried to tag each commend with a `#uid` for ease of referencing.There is so much cool stuff here that I feel like I could keep digging for a while longer, but will have to stop myself here:

  + all good, but I added something as a nice-to- have below!
  +  see below issue/bugs found in vignettes that causes them not to run properly
  +  see `Documentation - Functions` section
  +  see `Documentation - Examples` section

  + See `Functionality - Functionality` section below

Estimated hours spent reviewing: 16 - probably longer than it should have given it was my first review.

## Review Comments

- One thing I found challenging about this package is that its not straight forward to get help with
"?function_name" syntax. In `{rgee}` you are able to do it through `ee_help()` which is not as straight forward as ?function_name, but still better than nothing. Would there be a way to implement either ee_help() or ?function_name help here? #za_gc1

- It would be nice if the `How does it work` section of the readme explained a bit more about it's relation to `ee_extra`/`eemont`/`spectralIndex` python packages. #za_gc2
  + For most of this review I had thought you were just porting eemont functions over to rgeeExtra with reticulate, but at the end just realized that ee_extra (python) in the inst folder. Is that coming from: [ee_extra](https://github.com/r-earthengine/ee_extra) package somewhere in readme?

- It could be useful, especially for contributors, if the scope of the package was more clearly defined. The package naming makes it sounds like it could be a grab bag for all sorts of miscellaneous useful functions, but I don't think that is really the intent. Also as it is largely built in python, would you want to keep it that way or be open to more R development? I know that a lot of R development would be through `{rgee}` which wraps python, but it's still different than writing python to contribute to the `ee_extra` package - does that make sense? #za_gc3

- There are some nice defensive assertions you put in place on certain functions particularly the subletting functions. I think these really improve user experience and should be added more throughout the package. There are many cases where if the user inputs the wrong type of value the function will either a. eventually timeout, b. completely freeze/abort the session, c. give no warning and you only find out later. I've flagged a few in which some class assertions should help the user avoid this experience. #za_gc4

- `names()` - It seems likely you have already considered this, but I feel that it would make more sense if the output of `names()` was consistent across ImageCollection and Image classes. Currently names returns band names for Images and property names for ImageCollections. Why not just have `property_names()` and `names()` (or something like that) that works consistently since both Images an ImageCollections have both. #za_gc5

- would be nice to have a `list_spectral_indices()` type function (like eemont.listIndices())- would be cool if the object returned would provide the formula & reference for the index as well #za_gc6

- `Extra_closest()` - really cool function some comments (in addition to what I believe to be a small bug - see `Functionality-Functionality` section) #za_gc7

- Would it be possible to make it some functions allow the user to autocomplete arguments? I'm thinking of Extra_maskClouds, Extra_preprocess. #za_gc8

### Documentation - Installation instructions

- When I am using a more typical python workflow, I feel like I have a clear view of the venv and anything I'm adding to it. In this setup, it seems less visible. I believe that is largely intended and makes sense for a lot of R users. Nonetheless it would be nice if there was more documentation on this for people who are more comfortable with python. For example, I would personally find it useful to see an example of how to `pip install` all the necessary dependencies for both `{rgee}` and `{rgeeExtra}` into a pre-existing/or new `venv`. #za_di1

### Documentation - Examples

- No documentation on exported mean method https://github.com/r-earthengine/rgeeExtra/blob/7ea67d064e56e9429971d64440d8424df68f6249/R/ops-math.R#L258-L264
- example provided for `Summary.ee.image.Image` returns an image with 0 pixels so I am not reall ysure what exactly it is supposed to be doing? #za_de1
- Example on exported `names.ee.feature.Feature` method doesn't show function being used: https://github.com/r-earthengine/rgeeExtra/blob/7ea67d064e56e9429971d64440d8424df68f6249/R/ops-subsetting-fc.R#L57-L69 #za_de2
- Missing roxygen examples for `[` `[[` examples in `ops-subsetting-image.R`. It's not entirely clear what these are meant to do on an image so I think examples help #za_de3
- `ops-subsetting-imagecollection` the 2 top funcs have the same roxygen description but doing different things #za_de4
- `ops-subsetting.R` top example should have a `library(rgeeExtra)` call right? #za_de5
- see output of `devtools::check()` #za_de6

### Documentation - Functions

- `ee$Image$Extra_preprocess()` - would be nice to have more information on how to use in examples, and documentation/vignettes. It's not totally clear what the `...` arguments should be as it references `ee$Image$cloudmask()` which I do not believe is a function (maybe you meant `Extra_maskClouds`?). Since users should set parameters differently depending `img`/`ic` perhaps some explanation of default values should be given. Examples and vignettes sort of give impression that user should run without modifying parameters. I guess defaults are set for `Copernicus/S2_SR`? #za_df1
- `Extra_matchHistogram()`-  Would be useful to add `Map$addLayer()|Map$addLayer()` to roxygen example on function #za_df2
- `ops-math` - roxygen in below snippet says multiplication, but shows addition #za_df3
https://github.com/r-earthengine/rgeeExtra/blob/7ea67d064e56e9429971d64440d8424df68f6249/R/ops-math.R#L40-L44
- `ee_utils_gif_save()` roxygen says `@return No return value, called to write a GIF file.` but it does return a logical after writing the gif (i'm not sure why this is necessary) #za_df4
- `Extra_matchHistogram()` could be nice to provide a link and or some information on matching method #za_df5

### Documentation - Vignettes

- [aditional vignette](https://github.com/r-earthengine/rgeeExtra/blob/master/vignettes/Aditional.Rmd#L33)
  + Most examples are missing the `Extra_` prefix on the functions so they result in error #dv1
  + spelling: "Aditional" -> "Additional"  :-) #dv2
- I would check other Vignettes for this same issue. #dv3
- Could be nice to have a quick reminder at top of vignettes that in order to follow along users should install `{leaflet.extras2}` #dv4

## Functionality/ User Experience

### Functionality -Installation

- Worked fine for me. However in the past I have seen users struggle (including myself), but I believe this has gotten easier. #fi1

### Functionality - Functionality

- `Exra_Closest() very handy function! but some issues/questions
  + it returns and `ImageCollection` not an `Image` as written in the roxygen although, but based on how it is written to only return ImageCollection of 1 image, I think it should be converted to an Image? #za_ff1 (https://github.com/r-earthengine/rgeeExtra/blob/7ea67d064e56e9429971d64440d8424df68f6249/R/ee_ImageCollection.R#L18)
  + Can't take R formatted dates as written in roxygen (but should) #za_ff2

library(rgee)
library(rgeeExtra)
library(tidyverse)
ee_Initialize()
extra_Initialize()

url_s2 <- "COPERNICUS/S2"
ic <- ee$ImageCollection(url_s2)

date_chr <- "2023-09-10"
date_dt<- as_date(date_chr)
ic_closest_from_dt_obj <-  ic$Extra_closest(date_dt)
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named) :
#>   ee.ee_exception.EEException: Invalid argument specified for ee.Date(): 2023-09-10
#> Run `reticulate::py_last_error()` for details.
```
  + Could not find any documentation on what is supposed to happen on geographically large ImageCollection multiple images on same date. Theoretically as a user I might think that it would give me an ImageCollection containing all the images closes to that date, but I don't think that is what you were going for. Nonetheless I think we need some instructions in the roxygen and vignettes. Also this below cases is an example where `Extra_closest()` silently fails and then crashes/aborts the session on the next call that tries to do something with the resulting object. This is a case where some assertions would improve user experience. #za_ff3

ic_closest <-  ic$Extra_closest(date_chr) # works
class(ic_closest)
#> [1] "ee.imagecollection.ImageCollection" "ee.collection.Collection"           "ee.element.Element"
#>[4] "ee.computedobject.ComputedObject"   "ee.encodable.Encodable"             "python.builtin.object"

# if I run this in combination with above chunk my R session will freeze/abort
ee_get_date_ic(ic_closest)
```

  + **enhancement idea** - would be a nice feature if user could also supply an image instead of a date. The function then could find the closest image in the image collection to that date of the image supplied. Similar to functionalit of `find_closest()` function in this [example](https://developers.google.com/earth-engine/tutorials/community/histogram-matching). #za_ff4
- `Extra_minValue/Extra_maxValue()` - I think this function is only designed for `ee.Image` therefore some class assertions would be nice to avoid slow "ee.ee_exception.EEException: User memory limit exceeded" message if someone runs on anything else? #za_ff5
- `Extra_matchHistogram()` - probably a user error, but I had a hard time trying to reproduce the [Histogram Matching Example](https://developers.google.com/earth-engine/tutorials/community/histogram-matching) from Google. I think my issue likely relates to what was flagged in `#za_df1`  here is an attempt: #za_ff6
library(rgee)
library(rgeeExtra)
library(tidyverse)
ee_Initialize()
extra_Initialize()

# make geom
geometry = ee$Geometry$Polygon(
  list(
    c(-155.97117211519446, 20.09006980142336),
    c(-155.97117211519446, 19.7821681268256),
    c(-155.73256280122962, 19.7821681268256),
    c(-155.73256280122962, 20.09006980142336)
  )
)
skysat = ee$Image('SKYSAT/GEN-A/PUBLIC/ORTHO/RGB/s01_20161020T214047Z')$clip(geometry)
landsat_ic = ee$ImageCollection('LANDSAT/LC08/C01/T1_SR')$
  filterBounds(geometry) %>%
  # would be nice if vignettes showed good default presets for at least landsat & and sentinel 2
  ee$ImageCollection$Extra_preprocess()

date_skysat <- ee_get_date_img(skysat)$time_start # if you run Extra_closest w/ this it aborts session
date_skysat_chr<- as.character(as_date(date_skysat))

landsat_ic_closest <- ee$ImageCollection$Extra_closest(landsat_ic ,date_skysat_chr,32,"day")

# would be nice if Extra_closest returns img in the above
landsat_img_closest <- landsat_ic_closest[[1]] # but alas we have to convert it

landsat_img_rgb <- landsat_img_closest$select(
  opt_selectors = c( "B4","B3","B2"),
  opt_names = c("R","G","B")
)

bands <- list("R"="R","G"="G","B"="B")
matched <- ee$Image$Extra_matchHistogram(skysat,landsat_img_rgb,  bands)

vis_params_refl = list('min'=0, 'max'= 0.25)
vis_params_dn = list('min'= 0, 'max'= 255)
aoi_center_pt <- ee$Geometry$Point(c(-155.79584,19.9986))
Map$centerObject(aoi_center_pt,13)

m1 <- Map$addLayer(landsat_img_rgb, visParams = vis_params_refl,"landsat ref")
m2 <- Map$addLayer(skysat,visParams= vis_params_dn, 'SkySat source')
m1|m2 # looks good
m3 <- Map$addLayer(matched,visParams= vis_params_refl, 'SkySat matched')
m4 <- Map$addLayer(matched,visParams= vis_params_dn, 'SkySat matched')
m1|m3 # matches is just black
m1|m4 # matches is just black
```

For a user like me who wrote the code below what are some tools we have from `{rgee}` and `{rgeeExtra}` to help troubleshoot and figure out what is going on? This is all I can think of right now but doesn't get me much further? #za_ff7
# trying to trouble shoot - wrote this before I THINK i might have identified the problem
# in leaving the Extra_preprocess as default
rgeeExtra::ee_minValue(landsat_img_rgb)
rgeeExtra::ee_maxValue(landsat_img_rgb)
rgeeExtra::ee_minValue(skysat)
rgeeExtra::ee_maxValue(skysat)

rgeeExtra::ee_minValue(matched) # all 0s
rgeeExtra::ee_maxValue(matched) # all 0s
```

- `Extra_cloudMask()` would be nice to have this for ics (`ee_ImageCollection_maskClouds`) #za_ff8

### Functionality - Code style

- in vignettes and examples you seem to prefer the `%>%` syntax for chaining together `{rgee}`/`{rgeeExtra}` processes like this:

doi <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] %>%
  ee$Image$Extra_getDOI()
```

rather than:
doi <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]]$
  Extra_getDOI()

ic <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")

```

  + is this somehow best practices or you just think easier for R-users? I think the second looks cleaner (especially for longer chains) and might be easier for users migrating from the JavaScript code editor. #za_fc1
  + perhaps useful if it could be made clear somewhere in the documentation that you can do both styles and if there is a benefit to either of certain cases #za_fc2

- can remove leftover code
https://github.com/r-earthengine/rgeeExtra/blob/7ea67d064e56e9429971d64440d8424df68f6249/R/ops-math.R#L108-L110 #za_fc3
- In summary methods you rename the new band to "layer" no matter what operation( `summ_r$rename("layer")`)? Would it be safer to explicitly rename the band based on the band_name + Reducer name (i.e B1_max)? #za_fc4
- `raster.R` - why is this file called raster? name seems out of place? #za_fc5
- `ee_get()`- wonder if this name is a bit ambiguous considering all the other `get_*` functions there are in `rgee`,`GEE` and `rgeeExtra`... a couple ideas which make more sense to me filterIndex, filter_index, slice, subset, subset_index #za_fc6
- recommendation is use `<-` assignment over `->` -- #za_fc7 https://github.com/r-earthengine/rgeeExtra/blob/7ea67d064e56e9429971d64440d8424df68f6249/R/ee_gif.R#L187-L192
