### 2024-01-16 #608 @nmorandeira
Thanks [@csaybar](https://github.com/csaybar) for developing this excellent package!, and thanks @maurolepore for the invitation to review it. Sorry for the delay, I'm out of my office with very limited internet access. I provide some comments, mainly related to the documentation. As I said to @maurolepore, my views are as an end-user, since I have no experience in developing a package.

- [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor before starting your review).

- [x] A statement of need: clearly stating problems the software is designed to solve and its target audience in README.
_See below, I suggest adding more detail on this._

- [ ]  Installation instructions: for the development version of package and any non-standard dependencies in README.
_See my comment below. Installation instructions are insufficient._

- [x]  Vignette(s): demonstrating major functionality that runs successfully locally
_See my comments below._

- [ ]   Function Documentation: for all exported functions
_See my comments below._

- [x]   Examples: (that run successfully locally) for all exported functions
_See my comments below._

- [x]   Community guidelines: including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).

- [ ] Installation: Installation succeeds as documented.
I found problems initializing ee, see report below.

I haven't tested functionality, performance, automated tests and packaging guidelines.

Estimated hours spent reviewing: 10;  ~3 hours were spent in trying to solve the installation issues.

## Review Comments

+ In the Readme of the repository and in the rgeeExtra webpage, section "Why rgeeExtra is needed?"[https://github.com/r-earthengine/rgeeExtra?tab=readme-ov-file#why-rgeeextra-is-needed-] I would add a few sentences explaining the table, for the user to quickly know the extra features of the package in comparison with rgee.

For example, _"While GEE's primary language is JavaScript, which diverges from R's structure, rgeeExtra serves as a bridge"_. Isn't this true for rgee? What are the additional advantages of using rgeeExtra? I now know the answer :), after using the package and reviewing the Articles, but It would be nice for the user to quickly decide if it's worth installing rgeeExtra or no.

+ The equation provided as example is confusing because is called “ndvi” but includes a _**2_. I would name the object _“ndvi_squared”_, or alternatively remove the _** 2_ from the equation.

+ Add to the Readme.MD in the repository a link to https://r-earthengine.com/rgeeExtra/index.html
Articles are a nice vignette! I would highlight them.

+ Citation: is there an error? In [this section](https://r-earthengine.com/rgeeExtra/index.html#share-the-love-%EF%B8%8F), it says
```

‘citation("rgeeExtra")
To cite **rgee** in publications use:

  C Aybar, Q Wu, L Bautista, R Yali and A Barja (2020) rgee: An R
  package for interacting with Google Earth Engine Journal of Open
  Source Software URL https://github.com/r-spatial/rgee/.

A BibTeX entry for LaTeX users is

@Article{,
  title = {rgee: An R package for interacting with Google Earth Engine},
  author = {Cesar Aybar and Quisheng Wu and Lesly Bautista and Roy Yali and Antony Barja},
  journal = {Journal of Open Source Software},
}
‘
```

In R, running citation(“rgeeExtra”) gives the corrected text “To cite **rgeeExtra** in publications use”, but the same reference by Aybar et al. (2020) refering to rgee.

+ **Installation.** This was one of the main issues I encountered while reviewing the package. I was not able to initialize ee, here follows my error (probably related to miniconda or Python?). I couldn’t solve this issue in my laptop. Instead, I decided to conduct the review on other PC were I had already set up my environment. My suggestion is to include a clear step-to-step instruction to install the package and to initialize ee.

I followed the instructions provided here by Mauro: https://github.com/ropensci/software-review/issues/608#issuecomment-1755823040 .
When running `rgee::ee_install() ` I get this error:

I'm successfully using gee with Python in this laptop, but probably my R + Python environment is not correctly configured.

+  **Function Documentation:** after reviewing the Articles, it's not clear for me if all the functions included in rgeeExtra are listed in the examples. It seems to me that there may be extra functions that I have still not explored. A vignette including the list of functions and its parameters would be appreciated.

## Articles
Here I provide some comments on the articles.

### Introduction
SAVI computation. I get this error:

```
Error in py_get_attr_impl(x, name, silent) :
  AttributeError: type object 'Image' has no attribute 'spectralIndex'
```

Please check this error. I replaced “preprocess” with “Extra_preprocess”, and “spectralIndex” with “Extra_spectralIndex”. Next, I also replaced “geoviz_1” with “geoviz”. In this way I get no error, but I also don’t see any image in the visualizer.

Besides, if it’s possible to compute an spectral index this way, I would add a comment in the Readme file: after giving the ndvi or squared ndvi equation as an example of how an user can define a new formula using the rgeeExtra syntax, I would link to this SAVI example (or an NDVI example) as an additional advantage of the package, i.e. providing an easy way to compute well known spectral indexes.

### Image
Add dependency:
`library(leaflet.extras2)
`
**1. Arithmetic Operations.** The screenshot image shows global coverage, but the code gives a plot with zoom 9. Be consistent so that the user doesn't think he/she has done something wrong.

**2. Mathematical functions.** Comment that you can visualize each NDVI transformation by clicking on the layer selector.

**4. Tasseled Cap Transformation.**
I suggest replacing the display by the most used RGB combination for tasseled cap (R = TCB, G = TCG, B = TCW).

A reference to the citation for the the tasseled cap transformation would be nice, as well as for the other indexes included in the articles (NDVI, SAVI, etc).

**6. Histogram Matching**
The visualization parameters should be different for each image. A possible solution (I have also added the source scene for comparison purposes):

```
visParams_L8 <- list(
  bands = c("B4","B3","B2"),
  min = 0.1,
  max = 0.5,
  gamma = 1.5
)

visParams_L7 <- list(
  bands = c("B3","B2","B1"),
  min = 0.1,
  max = 0.5,
  gamma = 1.5
)

# Display matched and target images on the map for comparison.
Map$setCenter(lon = -121.984, lat = 47.847, zoom = 14)
m0 <- Map$addLayer(source, visParams = visParams_L8, "source")
m1 <- Map$addLayer(matched, visParams = visParams_L8, "matched")

m2 <- Map$addLayer(target, visParams = visParams_L7, "target")

m0 | m2

m1 | m2

```

**7. Cloud masking**
Please check the parameters. I don’t get the same display.

**8. Automated Preprocessing**
Please check the parameters. I don’t get the same display.

**9. Summary Methods.**
Add visualization instructions.

**10. Minimum and Maximum values.**
Requires geojsonio, which also has several dependencies. Add: `library(geojsonio)`
Also, the text says _“These functions offer two modes: “Rectangle” for using the image’s footprint and “Points” for sampling over it”_. Can you add examples for these two modes?

## Image Collection
**8. Complete GIF Toolkit for Earth Engine.**
Excellent tool! Please add the dependency:
`library(magick)`

Also, save the gif file to the working directory (It was saved in a temporal folder with a random name, so it’s difficult to find it).

**11. Get the temporal nearest image.**
Excellent tool! I would move this example before example (8). Also, the title is in italics, it seems to be an error.
As an user, I would find much more useful to find the closest cloud-free image.
