### 2024-01-18 #608 @csaybar
Thanks, I will start with the revision.

El mié, 17 ene 2024 a las 2:30, Mauro Lepore ***@***.***>)
escribió:

> Thanks a lot @nmorandeira <https://github.com/nmorandeira> for your
> review! Your perspective is very valuable and really nicely complements
> that of @zackarno <https://github.com/zackarno>.
>
> @csaybar <https://github.com/csaybar> you can now start addressing the
> reviews.
>
