### 2024-05-10 #608 @maurolepore
> If the author mentions changes might take time, [apply the holding label via typing @ropensci-review-bot put on hold](https://devdevguide.netlify.app/softwarereview_policies#policiesreviewprocess). You’ll get a reminder every 90 days (in the issue) to check in with the package author(s).

--https://devdevguide.netlify.app/softwarereview_editor

--

Hi @csaybar, shall we put this submission on hold?

It's totally fine to do so. The label helps us to better understand and manage the process.
