### 2024-05-13 #608 @csaybar
Hi everyone,

I sincerely apologize for my delayed response,

I had a bit of trouble properly solving all of your valid requests. As many of you know, rgee-extra is an extension of rgee. At the same time, rgee relies on the earthengine-api (Python). Therefore, updates to the Python client library directly impact rgee. Recently, the earthengine-api has undergone substantial changes concerning authentication (see updates [here](https://developers.google.com/earth-engine/changelog#python_api_v01383) and [here](https://developers.google.com/earth-engine/changelog#python_api_v01385)), which have break many of rgee’s functionalities.

After some internal discussions, we have decided to restructure rgee into a new package (to be named rgee2) to address installation & authentication problems.

@JulioContrerasH is working part-time on developing rgee2, but we do not expect it to be ready before the end of August. We are committed to keeping rgee-extra as an extension of rgee2. However, we prefer to pause further development on the project until rgee2 first version is released.

Thank you for your understanding, and sorry for the lengthy message. :(
Yes, @maurolepore, please can u move it to hold.
