### 2024-08-12 #608 @maurolepore
@csaybar

Is there any change?

The bot is reminding us to review the holding status. This is not to put any pressure but just to ensure we don't forget.

We'll get reminders every three months for a year, then we would need to move forward or close the submission.
