### 2025-01-25 #608 @csaybar
Hi @emilyriederer

Thank you so much for following up, I truly appreciate your understanding. I completely agree with the need to resolve this and regret that I haven’t been able to provide updates. Unfortunately, the changes required have proven more complex than anticipated, and with my energy stretched thin,  I no longer have the capacity to continue this work. We solved almost all the issues but the authentication problem persists.

Please feel free to close this out for now. My sincere apologies for the delays and the waste of time this has caused to everyone. I’m deeply grateful for all your patience and collaboration throughout this process.

Sorry everyone again and thank you for your kindness and flexibility.
