### 2025-01-26 #608 @emilyriederer
Hi @csaybar - thanks for the thoughtful note!

As  a matter of repo hygiene, I may close this for now, but please know we can re open at a later point in time if you want to complete the process.

Also, please let me say there is no reason to apologize. It sounds like the review process was hopefully still useful in its own way in surfacing some issues or ideas, and that's the more important outcome than a rubber stamp of approval.

As I have not been involved in this review, I'm not qualified to assess what issues are mission-critical versus nice-to-haves, but also do realize review feedback is not a one-way street. If you think there are design or technical reasons to not implement some of the suggestions, that is also a constructive discussion we could have here.

Up to you, going forward! Closing for now but we're still happy to engage on a path forward if it feels right later.
